Move context to helpers
